requirements.txt
Flask
Flask-CORS
requests
gunicorn
yfinances